Listão - separar o componente de período para usar em mais de um local
