+++ src/SME.SGP.WebClient/src/paginas/DiarioClasse/Listao/operacoes/listaoTabs/tabFrequencia/tabListaoFrequencia.js
@@ -1,17 +1,16 @@
 import { Col, Row } from 'antd';
 import _ from 'lodash';
-import React, { useCallback, useContext, useEffect, useState } from 'react';
+import React, { useCallback, useContext, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Alert, Loader, SelectComponent } from '~/componentes';
-import { SGP_SELECT_PERIODO_POR_COMPONENTE_CURRICULAR } from '~/componentes-sgp/filtro/idsCampos';
+import { Alert } from '~/componentes';
 import {
   setLimparModoEdicaoGeral,
   setTelaEmEdicao,
 } from '~/redux/modulos/geral/actions';
 import { erros } from '~/servicos';
-import ServicoPeriodoEscolar from '~/servicos/Paginas/Calendario/ServicoPeriodoEscolar';
 import ServicoFrequencia from '~/servicos/Paginas/DiarioClasse/ServicoFrequencia';
 import ListaoContext from '../../../listaoContext';
+import PeriodoEscolarListao from '../componentes/periodoEscolarListao';
 import ListaoAuditoria from './lista/componentes/listaoAuditoria';
 import ListaoListaFrequencia from './lista/listaoListaFrequencia';
 
@@ -19,19 +18,13 @@ const TabListaoFrequencia = () => {
   const dispatch = useDispatch();
 
   const usuario = useSelector(store => store.usuario);
-  const telaEmEdicao = useSelector(store => store.geral.telaEmEdicao);
   const { turmaSelecionada } = usuario;
-  const { turma } = turmaSelecionada;
-  const acaoTelaEmEdicao = useSelector(store => store.geral.acaoTelaEmEdicao);
 
   const {
     componenteCurricular,
     bimestreOperacoes,
     setExibirLoaderGeral,
-    listaPeriodos,
-    setListaPeriodos,
     periodo,
-    setPeriodo,
     dadosFrequencia,
     setDadosFrequencia,
     setListaTiposFrequencia,
@@ -39,102 +32,25 @@ const TabListaoFrequencia = () => {
     periodoAbertoListao,
   } = useContext(ListaoContext);
 
-  const [exibirLoaderPeriodo, setExibirLoaderPeriodo] = useState(false);
-
-  const desabilitarPeriodo =
-    !turma ||
-    !bimestreOperacoes ||
-    !componenteCurricular?.codigoComponenteCurricular ||
-    listaPeriodos?.length === 1 ||
-    !listaPeriodos?.length;
-
   const limparFrequencia = () => {
-    setDadosIniciaisFrequencia({});
-    setDadosFrequencia({});
+    setDadosIniciaisFrequencia();
+    setDadosFrequencia();
   };
 
   useEffect(() => {
     return () => {
       limparFrequencia();
-      setPeriodo();
-      setListaPeriodos([]);
+      setListaTiposFrequencia([]);
+      dispatch(setLimparModoEdicaoGeral(false));
     };
   }, []);
-
-  const obterPeriodoPorComponente = useCallback(async () => {
-    limparFrequencia();
-    setExibirLoaderPeriodo(true);
-    const resposta = await ServicoPeriodoEscolar.obterPeriodoPorComponente(
-      turma,
-      componenteCurricular?.codigoComponenteCurricular,
-      componenteCurricular?.regencia,
-      bimestreOperacoes
-    )
-      .catch(e => erros(e))
-      .finally(() => setExibirLoaderPeriodo(false));
-
-    if (resposta?.data?.length) {
-      const lista = resposta.data.map(item => {
-        return { ...item, id: String(item.id) };
-      });
-      setListaPeriodos(lista);
-      if (lista.length === 1) {
-        setPeriodo(lista[0]);
-      }
-    } else {
-      setListaPeriodos([]);
-    }
-  }, [componenteCurricular, turma, bimestreOperacoes]);
-
-  useEffect(() => {
-    setPeriodo();
-    setListaPeriodos([]);
-    if (componenteCurricular?.codigoComponenteCurricular && bimestreOperacoes) {
-      obterPeriodoPorComponente();
-    } else {
-      limparFrequencia();
-    }
-  }, [bimestreOperacoes]);
-
-  const obterPeriodoSelecionado = id => {
-    if (id && listaPeriodos?.length) {
-      let periodoSelecionado = null;
-
-      periodoSelecionado = listaPeriodos?.find(
-        item => Number(item?.id) === Number(id)
-      );
-      return periodoSelecionado;
-    }
-    return null;
-  };
-
-  const setarPeriodo = valor => {
-    const periodoSelecionado = obterPeriodoSelecionado(valor);
-    if (periodoSelecionado) {
-      setPeriodo({ ...periodoSelecionado });
-    } else {
-      setPeriodo();
-    }
-  };
-
-  const onChangePeriodo = async valor => {
-    if (telaEmEdicao) {
-      const salvou = await acaoTelaEmEdicao();
-      if (salvou) {
-        limparFrequencia();
-        setarPeriodo(valor);
-      }
-    } else {
-      setarPeriodo(valor);
-    }
-  };
 
   const obterFrequenciasPorPeriodo = useCallback(async () => {
     setExibirLoaderGeral(true);
     const resposta = await ServicoFrequencia.obterFrequenciasPorPeriodo(
       periodo?.dataInicio,
       periodo?.dataFim,
-      turma,
+      turmaSelecionada?.turma,
       componenteCurricular?.codigoComponenteCurricular,
       componenteCurricular?.id
     )
@@ -173,10 +89,8 @@ const TabListaoFrequencia = () => {
     dispatch,
     periodoAbertoListao,
     componenteCurricular,
-    turma,
     turmaSelecionada,
     periodo,
-    setExibirLoaderGeral,
   ]);
 
   useEffect(() => {
@@ -185,23 +99,12 @@ const TabListaoFrequencia = () => {
       periodo?.dataInicio &&
       periodo?.dataFim &&
       componenteCurricular?.codigoComponenteCurricular &&
-      turma &&
+      turmaSelecionada?.turma &&
       bimestreOperacoes
     ) {
       obterFrequenciasPorPeriodo();
     }
   }, [periodo]);
-
-  useEffect(() => {
-    return () => {
-      setListaPeriodos([]);
-      setPeriodo();
-      setListaTiposFrequencia([]);
-      setDadosFrequencia();
-      setDadosIniciaisFrequencia();
-      dispatch(setLimparModoEdicaoGeral(false));
-    };
-  }, []);
 
   return (
     <>
@@ -219,22 +122,10 @@ const TabListaoFrequencia = () => {
       ) : (
         <></>
       )}
+
       <Row gutter={[24, 24]}>
         <Col sm={24} md={12} lg={8}>
-          <Loader loading={exibirLoaderPeriodo} ignorarTip>
-            <SelectComponent
-              label="Período"
-              id={SGP_SELECT_PERIODO_POR_COMPONENTE_CURRICULAR}
-              lista={listaPeriodos}
-              valueOption="id"
-              valueText="periodoEscolar"
-              valueSelect={periodo?.id}
-              onChange={onChangePeriodo}
-              placeholder="Selecione um Período"
-              showSearch
-              disabled={desabilitarPeriodo}
-            />
-          </Loader>
+          <PeriodoEscolarListao limparDadosTabSelecionada={limparFrequencia} />
         </Col>
       </Row>
